move files from other directory to LWA_EPIC, clean up imports
